requirements.txt
appdirs==1.4.4
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.1.7
dnspython==2.6.1
flexcache==0.3
flexparser==0.3.1
joblib==1.4.2
mathparse==0.1.2
nltk==3.8.1
Pint==0.24.1
pymongo==3.13.0
python-dateutil==2.7.5
python-dotenv==1.0.1
pytz==2024.1
PyYAML==3.13
regex==2024.5.15
six==1.16.0
SQLAlchemy==1.2.19
tqdm==4.66.4
typing_extensions==4.12.2
